[MOVE] API 컨벤션에 맞게 파일 경로 이동
