+++ src/lib/apis/auth/auth.ts
@@ -1,12 +1,8 @@
-import { api } from "@/api/client";
-
 export type SocialProvider = "kakao" | "naver" | "google";
 
+// 소셜 로그인
 export const loginWithSocial = (provider: SocialProvider) => {
   const env = process.env.NODE_ENV === "development" ? "local" : "production";
   const baseUrl = process.env.NEXT_PUBLIC_API_BASE_URL ?? "";
   window.location.assign(`${baseUrl}/oauth2/authorization/${provider}?env=${env}`);
 };
-
-export const postDeviceToken = (pushToken: string) =>
-  api.post("/api/users/me/device-tokens", { platform: "WEB", pushToken });
