--- src/lib/apis/user/getOnboardingStatus.ts
@@ -1,4 +1,0 @@
-import { api } from "@/api/client";
-import type { OnboardingStatus } from "@/types/user/onboarding";
-
-export const getOnboardingStatus = () => api.get<OnboardingStatus>("/api/onboarding/status");
