+++ src/lib/apis/user/user.ts
@@ -1,4 +1,5 @@
 import { serverApi } from "@/api/server";
 import type { UserProfile } from "@/types/user/user";
 
+// 내 정보 조회
 export const getMe = () => serverApi.get<UserProfile>("/api/users/me");
