+++ src/lib/hooks/user/usePostOnboardingComplete.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { postOnboardingComplete } from "@/lib/apis/user/postOnboardingComplete";
+import { postOnboardingComplete } from "@/lib/apis/user/onboarding";
 import type { OnboardingRequest } from "@/types/user/onboarding";
 
 export const usePostOnboardingComplete = () =>
